설비수정 완료

+++ FamTec/FamTec/FamTec/Server/Controllers/Facility/NetworkFacilityController.cs
@@ -34,6 +34,15 @@ namespace FamTec.Server.Controllers.Facility
                 if (HttpContext is null)
                     return BadRequest();
 
+                if (String.IsNullOrWhiteSpace(dto.Category))
+                    return NoContent();
+
+                if (String.IsNullOrWhiteSpace(dto.Name))
+                    return NoContent();
+
+                if (dto.RoomTbId is null)
+                    return NoContent();
+
                 if (files is not null)
                 {
                     if (files.Length > Common.MEGABYTE_1)
@@ -86,7 +95,7 @@ namespace FamTec.Server.Controllers.Facility
                 ResponseList<FacilityListDTO>? model = await NetworkFacilityService.GetNetworkFacilityListService(HttpContext);
 
                 if (model is null)
-                    return BadRequest();
+                    return BadRequest(); 
 
                 if (model.code == 200)
                     return Ok(model);
@@ -110,7 +119,7 @@ namespace FamTec.Server.Controllers.Facility
                 if (HttpContext is null)
                     return BadRequest();
 
-                ResponseUnit<FacilityDetailDTO?> model = await NetworkFacilityService.GetNetworkDetailFacilityService(HttpContext, facilityid);
+                ResponseUnit<FacilityDetailDTO> model = await NetworkFacilityService.GetNetworkDetailFacilityService(HttpContext, facilityid);
                 if (model is null)
                     return BadRequest();
 
@@ -135,6 +144,18 @@ namespace FamTec.Server.Controllers.Facility
             {
                 if (HttpContext is null)
                     return BadRequest();
+
+                if (dto.ID is null)
+                    return NoContent();
+
+                if (String.IsNullOrWhiteSpace(dto.Category))
+                    return NoContent();
+
+                if (String.IsNullOrWhiteSpace(dto.Name))
+                    return NoContent();
+
+                if (dto.RoomTbId is null)
+                    return NoContent();
 
                 if (files is not null)
                 {
@@ -183,6 +204,12 @@ namespace FamTec.Server.Controllers.Facility
             {
                 if (HttpContext is null)
                     return BadRequest();
+
+                if (delIdx is null)
+                    return NoContent();
+
+                if (delIdx.Count() == 0)
+                    return NoContent();
 
                 ResponseUnit<int?> model = await NetworkFacilityService.DeleteNetworkFacilityService(HttpContext, delIdx);
                 if (model is null)
